+++ packages/tests/services/PointerService.spec.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, afterEach } from 'vitest';
 import { PointerServiceProps, usePointer } from '@studiometa/js-toolkit';
-import { createEvent } from '#test-utils';
+import { h, dispatch } from '#test-utils';
 
 const key = 'usePointer';
 const { add, remove, props } = usePointer();
@@ -30,22 +30,31 @@ describe('usePointer', () => {
     expect(typeof props).toBe('function');
   });
 
+  it('should support using a target element', () => {
+    const div = h('div');
+    const fn = vi.fn();
+    const service = usePointer(div);
+    service.add('key', fn);
+    dispatch(document.documentElement, 'mouseenter', { clientX: 0, clientY: 0 });
+    expect(fn).toHaveBeenLastCalledWith(service.props());
+  });
+
   it('should trigger the callbacks on mousedown and mouseup', () => {
     const { fn, pointerProps } = getContext();
-    document.dispatchEvent(createEvent('mousedown'));
+    dispatch(document, 'mousedown');
     expect(fn).toHaveBeenCalledTimes(1);
     expect(pointerProps().isDown).toBe(true);
-    document.dispatchEvent(createEvent('mouseup'));
+    dispatch(document, 'mouseup');
     expect(fn).toHaveBeenCalledTimes(2);
     expect(pointerProps().isDown).toBe(false);
   });
 
   it('should trigger the callbacks on touchstart and touchend', () => {
     const { fn, pointerProps } = getContext();
-    document.dispatchEvent(new TouchEvent('touchstart'));
+    dispatch(document, 'touchstart');
     expect(fn).toHaveBeenCalledTimes(1);
     expect(pointerProps().isDown).toBe(true);
-    document.dispatchEvent(new TouchEvent('touchend'));
+    dispatch(document, 'touchend');
     expect(fn).toHaveBeenCalledTimes(2);
     expect(pointerProps().isDown).toBe(false);
   });
@@ -54,7 +63,7 @@ describe('usePointer', () => {
     const { fn } = getContext();
     const otherFn = vi.fn();
     add('otherUsePointer', otherFn);
-    document.dispatchEvent(createEvent('mouseup'));
+    dispatch(document, 'mouseup');
     expect(fn).toHaveBeenCalledTimes(1);
     expect(otherFn).toHaveBeenCalledTimes(1);
     remove('otherUsePointer');
@@ -65,42 +74,40 @@ describe('usePointer', () => {
     remove('usePointer');
     const options = { clientX: 0, clientY: 0 };
     const touchOptions = { touches: [{ clientX: 0, clientY: 0 }] };
-    document.dispatchEvent(createEvent('mouseup', options));
-    document.dispatchEvent(createEvent('mousedown', options));
-    document.dispatchEvent(createEvent('mousemove', options));
-    document.dispatchEvent(createEvent('touchstart', touchOptions));
-    document.dispatchEvent(createEvent('touchend', touchOptions));
-    document.dispatchEvent(createEvent('touchmove', touchOptions));
+    dispatch(document, 'mouseup', options);
+    dispatch(document, 'mousedown', options);
+    dispatch(document, 'mousemove', options);
+    dispatch(document, 'touchstart', touchOptions);
+    dispatch(document, 'touchend', touchOptions);
+    dispatch(document, 'touchmove', touchOptions);
     expect(fn).toHaveBeenCalledTimes(0);
   });
 
   it('should trigger the callbacks on mousemove', () => {
     const { fn } = getContext();
-    document.dispatchEvent(createEvent('mousemove', { clientX: 0, clientY: 0 }));
+    dispatch(document, 'mousemove', { clientX: 0, clientY: 0 });
     const progress = 0.1;
     const x = Math.round(window.innerWidth * progress);
     const y = Math.round(window.innerHeight * progress);
 
-    const event = createEvent('mousemove', { clientX: x, clientY: y });
-    document.dispatchEvent(event);
+    dispatch(document, 'mousemove', { clientX: x, clientY: y });
 
     expect(fn).toHaveBeenLastCalledWith(props());
 
     const newX = x + 10;
-    document.dispatchEvent(createEvent('mousemove', { clientX: newX, clientY: y }));
+    dispatch(document, 'mousemove', { clientX: newX, clientY: y });
     expect(fn).toHaveBeenLastCalledWith(props());
   });
 
   it('should trigger the callbacks on touchmove', () => {
     const { fn } = getContext();
-    document.dispatchEvent(createEvent('touchmove', { touches: [{ clientX: 0, clientY: 0 }] }));
+    dispatch(document, 'touchmove', { touches: [{ clientX: 0, clientY: 0 }] });
 
     const progress = 0.1;
     const x = Math.round(window.innerWidth * progress);
     const y = Math.round(window.innerHeight * progress);
 
-    const event = createEvent('touchmove', { touches: [{ clientX: x, clientY: y }] });
-    document.dispatchEvent(event);
+    const event = dispatch(document, 'touchmove', { touches: [{ clientX: x, clientY: y }] });
 
     expect(fn).toHaveBeenLastCalledWith({
       event,
@@ -130,8 +137,9 @@ describe('usePointer', () => {
     });
 
     const newX = x + 10;
-    const otherEvent = createEvent('touchmove', { touches: [{ clientX: newX, clientY: y }] });
-    document.dispatchEvent(otherEvent);
+    const otherEvent = dispatch(document, 'touchmove', {
+      touches: [{ clientX: newX, clientY: y }],
+    });
     expect(fn).toHaveBeenLastCalledWith({
       event: otherEvent,
       isDown: false,